backend/requirements.txt
aiofiles==25.1.0
alembic==1.12.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.1.2
billiard==4.2.4
boto3==1.28.56
botocore==1.31.85
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.0.0
fastapi==0.123.5
greenlet==3.2.4
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
kombu==5.6.1
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.12
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pytest==7.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
qrcode==8.2
redis==5.0.1
requests==2.32.5
resend==2.19.0
rsa==4.9.1
s3transfer==0.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.0.7
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
